Split the cache classes out to their own files
